+++ velox/core/PlanNode.h
@@ -18,7 +18,7 @@
 #include "velox/connectors/Connector.h"
 #include "velox/core/Expressions.h"
 
-#include "velox/vector/arrow/Bridge.h"
+#include "velox/vector/arrow/c/Bridge.h"
 
 namespace facebook::velox::core {
