+++ velox/vector/arrow/c/tests/ArrowBridgeSchemaTest.cpp
@@ -17,7 +17,7 @@
 #include "gtest/gtest.h"
 
 #include "velox/common/base/Nulls.h"
-#include "velox/vector/arrow/Bridge.h"
+#include "velox/vector/arrow/c/Bridge.h"
 
 namespace {
